API/auto_api/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from . import views

router = DefaultRouter()
router.register(r'cars', views.CarViewSet)
router.register(r'services', views.ServiceViewSet)
router.register(r'orders', views.OrderViewSet)
router.register(r'contact', views.ContactMessageView)

urlpatterns = [
    path('', include(router.urls)),
    path('register/', views.RegisterView.as_view(), name='register'),

]