API/requirements.txt
Django>=5.0,<5.1
djangorestframework
djangorestframework-simplejwt
drf-yasg
django-filter
psycopg2-binary  # или оставь без БД, используй SQLite для учебного проекта
Pillow
openpyxl
django-cors-headers
django-import-export